(compat) Added supported features and generation across the Runtime / DataStore boundary
