+++ packages/runtime/container-runtime/src/runtimeLayerCompatState.ts
@@ -1,0 +1,143 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import {
+	checkLayerCompatibility,
+	type ILayerCompatDetails,
+	type ILayerCompatSupportRequirements,
+} from "@fluid-internal/client-utils";
+import type { ICriticalContainerError } from "@fluidframework/container-definitions";
+import { UsageError } from "@fluidframework/telemetry-utils/internal";
+
+import { pkgVersion } from "./packageVersion.js";
+
+/**
+ * The core compatibility details of the Runtime layer that is the same across all layer boundaries.
+ * @internal
+ */
+export const RuntimeCoreCompatDetails = {
+	/**
+	 * The package version of the Runtime layer.
+	 */
+	pkgVersion,
+	/**
+	 * The current generation of the Runtime layer.
+	 */
+	generation: 1,
+};
+
+/**
+ * Runtime's compatibility details that is exposed to the Loader layer.
+ * @internal
+ */
+export const RuntimeCompatDetailsForLoader: ILayerCompatDetails = {
+	...RuntimeCoreCompatDetails,
+	/**
+	 * The features supported by the Runtime layer across the Runtime / Loader boundary.
+	 */
+	supportedFeatures: new Set<string>(),
+};
+
+/**
+ * The requirements that the Loader layer must meet to be compatible with this Runtime.
+ * @internal
+ */
+export const LoaderSupportRequirements: ILayerCompatSupportRequirements = {
+	/**
+	 * Minimum generation that Loader must be at to be compatible with Runtime. Note that 0 is used here so
+	 * that Loader layers before the introduction of the layer compatibility enforcement are compatible.
+	 */
+	minSupportedGeneration: 0,
+	/**
+	 * The features that the Loader must support to be compatible with Runtime.
+	 */
+	requiredFeatures: [],
+};
+
+/**
+ * Runtime's compatibility details that is exposed to the DataStore layer.
+ * @internal
+ */
+export const RuntimeCompatDetailsForDataStore: ILayerCompatDetails = {
+	...RuntimeCoreCompatDetails,
+	/**
+	 * The features supported by the Runtime layer across the Runtime / Loader boundary.
+	 */
+	supportedFeatures: new Set<string>(),
+};
+
+/**
+ * The requirements that the DataStore layer must meet to be compatible with this Runtime.
+ * @internal
+ */
+export const DataStoreSupportRequirements: ILayerCompatSupportRequirements = {
+	/**
+	 * Minimum generation that DataStore must be at to be compatible with Runtime. Note that 0 is used here so
+	 * that DataStore layers before the introduction of the layer compatibility enforcement are compatible.
+	 */
+	minSupportedGeneration: 0,
+	/**
+	 * The features that the DataStore must support to be compatible with Runtime.
+	 */
+	requiredFeatures: [],
+};
+
+/**
+ * Validates that the Loader layer is compatible with this Runtime.
+ * @internal
+ */
+export function validateLoaderCompatibility(
+	maybeLoaderCompatDetails: ILayerCompatDetails | undefined,
+	disposeFn: (error?: ICriticalContainerError) => void,
+): void {
+	const layerCheckResult = checkLayerCompatibility(
+		LoaderSupportRequirements,
+		maybeLoaderCompatDetails,
+	);
+	if (!layerCheckResult.isCompatible) {
+		const error = new UsageError("Runtime is not compatible with Loader", {
+			errorDetails: JSON.stringify({
+				runtimeVersion: RuntimeCoreCompatDetails.pkgVersion,
+				loaderVersion: maybeLoaderCompatDetails?.pkgVersion,
+				runtimeGeneration: RuntimeCoreCompatDetails.generation,
+				loaderGeneration: maybeLoaderCompatDetails?.generation,
+				minSupportedGeneration: LoaderSupportRequirements.minSupportedGeneration,
+				isGenerationCompatible: layerCheckResult.isGenerationCompatible,
+				unsupportedFeatures: layerCheckResult.unsupportedFeatures,
+			}),
+		});
+		disposeFn(error);
+		throw error;
+	}
+}
+
+/**
+ * Validates that the DataStore layer is compatible with this Runtime.
+ * @internal
+ */
+export function validateDatastoreCompatibility(
+	maybeDataStoreCompatDetails: ILayerCompatDetails | undefined,
+	disposeFn: () => void,
+): void {
+	const layerCheckResult = checkLayerCompatibility(
+		DataStoreSupportRequirements,
+		maybeDataStoreCompatDetails,
+	);
+	if (!layerCheckResult.isCompatible) {
+		const error = new UsageError("Runtime is not compatible with DataStore", {
+			errorDetails: JSON.stringify({
+				runtimeVersion: RuntimeCoreCompatDetails.pkgVersion,
+				dataStoreVersion: maybeDataStoreCompatDetails?.pkgVersion,
+				runtimeGeneration: RuntimeCoreCompatDetails.generation,
+				dataStoreGeneration: maybeDataStoreCompatDetails?.generation,
+				minSupportedGeneration: DataStoreSupportRequirements.minSupportedGeneration,
+				isGenerationCompatible: layerCheckResult.isGenerationCompatible,
+				unsupportedFeatures: layerCheckResult.unsupportedFeatures,
+			}),
+		});
+		disposeFn();
+		throw error;
+	}
+}
