+++ packages/runtime/container-runtime/src/test/dataStoreContext.spec.ts
@@ -10,7 +10,6 @@ import { AttachState } from "@fluidframework/container-definitions";
 import { ContainerErrorTypes } from "@fluidframework/container-definitions/internal";
 import {
 	FluidObject,
-	ITelemetryBaseLogger,
 	Tagged,
 	TelemetryBaseEventPropertyType,
 } from "@fluidframework/core-interfaces";
@@ -69,6 +68,11 @@ import {
 	summarizerClientType,
 } from "../summary/index.js";
 
+import {
+	createParentContext,
+	createSummarizerNodeAndGetCreateFn,
+} from "./dataStoreCreationHelper.js";
+
 describe("Data Store Context Tests", () => {
 	const dataStoreId = "Test1";
 	const emptyGCData: IGarbageCollectionData = { gcNodes: {} };
@@ -76,59 +80,15 @@ describe("Data Store Context Tests", () => {
 
 	describe("LocalFluidDataStoreContext", () => {
 		let localDataStoreContext: LocalFluidDataStoreContext;
-		let storage: IDocumentStorageService;
-		let scope: FluidObject;
+		const storage = {} as unknown as IDocumentStorageService;
+		const scope = {} as unknown as FluidObject;
 		const makeLocallyVisibleFn = () => {};
 		let parentContext: IFluidParentContext;
 		let summarizerNode: IRootSummarizerNodeWithGC;
 
-		function createParentContext(
-			logger: ITelemetryBaseLogger = createChildLogger(),
-			clientDetails = {} as unknown as IFluidParentContext["clientDetails"],
-			submitMessage: IFluidParentContext["submitMessage"] = () => {},
-		): IFluidParentContext {
-			const factory: IFluidDataStoreFactory = {
-				type: "store-type",
-				get IFluidDataStoreFactory() {
-					return factory;
-				},
-				instantiateDataStore: async (context: IFluidDataStoreContext) =>
-					new MockFluidDataStoreRuntime(),
-			};
-			const registry: IFluidDataStoreRegistry = {
-				get IFluidDataStoreRegistry() {
-					return registry;
-				},
-				get: async (pkg) => (pkg === "BOGUS" ? undefined : factory),
-			};
-			return {
-				IFluidDataStoreRegistry: registry,
-				baseLogger: logger,
-				clientDetails,
-				submitMessage,
-			} satisfies Partial<IFluidParentContext> as unknown as IFluidParentContext;
-		}
-
 		beforeEach(async () => {
-			summarizerNode = createRootSummarizerNodeWithGC(
-				createChildLogger(),
-				(() => undefined) as unknown as SummarizeInternalFn,
-				0,
-				0,
-			);
-			summarizerNode.startSummary(0, createChildLogger(), 0);
-
-			createSummarizerNodeFn = (
-				summarizeInternal: SummarizeInternalFn,
-				getGCDataFn: () => Promise<IGarbageCollectionData>,
-			) =>
-				summarizerNode.createChild(
-					summarizeInternal,
-					dataStoreId,
-					{ type: CreateSummarizerNodeSource.Local },
-					undefined,
-					getGCDataFn,
-				);
+			({ summarizerNode, createSummarizerNodeFn } =
+				createSummarizerNodeAndGetCreateFn(dataStoreId));
 			parentContext = createParentContext();
 		});
 
@@ -557,7 +517,7 @@ describe("Data Store Context Tests", () => {
 		let remoteDataStoreContext: RemoteFluidDataStoreContext;
 		let dataStoreAttributes: ReadFluidDataStoreAttributes;
 		const storage: Partial<IDocumentStorageService> = {};
-		let scope: FluidObject;
+		const scope = {} as unknown as FluidObject;
 		let summarizerNode: IRootSummarizerNodeWithGC;
 		let parentContext: IFluidParentContext;
 
@@ -1013,8 +973,8 @@ describe("Data Store Context Tests", () => {
 
 	describe("LocalDetachedFluidDataStoreContext", () => {
 		let localDataStoreContext: LocalDetachedFluidDataStoreContext;
-		let storage: IDocumentStorageService;
-		let scope: FluidObject;
+		const storage = {} as unknown as IDocumentStorageService;
+		const scope = {} as unknown as FluidObject;
 		let factory: IFluidDataStoreFactory;
 		const makeLocallyVisibleFn = () => {};
 		const channelToDataStoreFn = (fluidDataStore: IFluidDataStoreChannel) =>
