+++ packages/runtime/test-runtime-utils/api-report/test-runtime-utils.legacy.alpha.api.md
@@ -339,6 +339,8 @@ export class MockFluidDataStoreContext implements IFluidDataStoreContext {
     // (undocumented)
     IFluidHandleContext: IFluidHandleContext;
     // (undocumented)
+    ILayerCompatDetails?: unknown;
+    // (undocumented)
     isLocalDataStore: boolean;
     // (undocumented)
     makeLocallyVisible(): void;
@@ -441,6 +443,8 @@ export class MockFluidDataStoreRuntime extends EventEmitter implements IFluidDat
     idCompressor: (IIdCompressor & IIdCompressorCore) | undefined;
     // (undocumented)
     get IFluidHandleContext(): IFluidHandleContext;
+    // (undocumented)
+    ILayerCompatDetails?: unknown;
     // (undocumented)
     get isAttached(): boolean;
     // (undocumented)
